fix: key chunk maps on integer coordinates instead of float vec2
Float-keyed glm::vec2 hashing loses precision far from origin,
producing duplicate/garbled chunk keys. Switch chunk maps, the
processing set, frustum culling and mesh generation to glm::ivec2
chunk coordinates.

Closes #35

+++ src/Frustum.h
@@ -4,7 +4,7 @@
 class Frustum {
  public:
   Frustum(const Camera *camera);
-  bool isChunkInside(const glm::vec2 &position);
+  bool isChunkInside(const glm::ivec2 &position);
 
   Plane m_NearFace;
   Plane m_FarFace;
